variable type has to be double

+++ 0x03-debugging/2-largest_number.c
@@ -8,9 +8,9 @@
  * Return: largest number
  */
 
-int largest_number(int a, int b, int c)
+double largest_number(double a, double b, double c)
 {
-	int largest;
+	double largest;
 
 	if (c >= a && c >= b)
 	{
